Ready to Deploy fix some issue and its stable for production ...
